requirements (4).txt
flask==3.0.0
anthropic==0.49.0
sendgrid==6.11.0
gunicorn==21.2.0
stripe==7.12.0
psycopg2-binary==2.9.9
pdfplumber==0.11.0
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
weasyprint==62.3
requests==2.31.0
